+++ tests/testthat/test-workStudy.R
@@ -53,6 +53,7 @@ test_that('Test working with AcousticStudy object', {
         is.na(ancillary(exData[[1]])$ici[[2]]$ici)
     ))
     expect_true(all(c('Click_Detector_1_ici', 'All_ici') %in% names(ancillary(exData[[1]])$measures)))
+    expect_true(all(c('Click_Detector_1_ici', 'All_ici') %in% names(getMeasures(exData[[1]]))))
     expect_true(all(c('Click_Detector_1_ici', 'All_ici') %in% names(getClickData(exData[[1]]))))
     iciData <- getICI(exData, 'data')
     expect_true(all(c('Click_Detector_1', 'All') %in% names(iciData[[1]])))