+++ NEWS.md
@@ -5,6 +5,8 @@
 * `calculateAverageSpectra` would have problems when doing multiple events that had detections
 with repeated UIDs
 
+* `matchEnvData` now has a `depth` parameter to specify depth or range of depths to match
+
 ## PAMpal 0.15.1
 
 * Added ability to add functions during `PAMpalSettings` call